Added:    `useHostedVideos` custom hook to encapsulate hosted video's duration fetching logic (src/hooks/useHostedVideos.tsx)
Refactored:   Simplified `<HostedVideoCard />` with imported custom hook (src/components/videos/hosted/HostedVideoCard.tsx)
